remove monomorphic optics
