--- src/main/scala/optics/mono/Lens.scala
@@ -1,6 +1,0 @@
-package optics.mono
-
-object Lens {
-  def apply[From, To](_get: From => To, _replace: To => From => From): Lens[From, To] =
-    EOptional(from => Right(_get(from)), _replace)
-}
